+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.2 (unreleased)
 ================
 
+- Expand benchmarks to include ang2pix, nest2ring and ring2nest. [#62]
+
 - Use OpenMP to parallelize the Cython wrappers. [#59]
 
 0.1 (2017-10-01)